refactor after merge and ui fixes
