+++ src/components/admin_panel/UserManager.js
@@ -1,0 +1,14 @@
+export default class UserManager {
+
+    removeUser(user) {
+        //TODO add logic
+    }
+
+    addUser(user) {
+        //TODO add logic
+    }
+
+    updateUser() {
+        //TODO add logic
+    }
+}
