+++ src/components/styles/admin-panel-styles.css
@@ -5,13 +5,33 @@
     margin: 0rem auto auto auto;
 }
 
+.card:hover {
+    box-shadow: 2px 4px 8px 0px rgb(0, 0, 0);
+    transform: scale(1.0);
+}
+
+.modal-500w {
+    width: 470px;
+}
+
+.defaultButtonRadius {
+    border-radius: 12px;
+}
+
 .adminPanelHeader {
     padding: 6px;
     padding-left: 20px;
+    padding-top: 15px;
     background-color: #4169E1;
     color: aliceblue;
-    max-width: 100%;
+    width: 280px;
     text-align: left;
+    border-radius: 0px 0px 40px 0px;
+}
+
+.kurwaaa {
+    max-width: 100%;
+    background-color: aquamarine;
 }
 
 .listItemBody {
